Got Project 1 working between title and explore states. Fixed minor warnings. Added cursor setting to TUI init function. Added overlapping windows UI tests and a mock main menu. Added default constructor to ExplorationProcessor, and MapRepository. Added load/unload methods to GameStates (may have been unnecessary).

+++ PDCurseControls/src/TextField.cpp
@@ -108,7 +108,7 @@ bool TextField::inputChar(int c)
 		cursorPos = 0;
 		break;
 	case KEY_END: 
-		cursorPos = text.length();
+		cursorPos = (unsigned short)text.length();
 		break;
 	case KEY_MOUSE: //does nothing right now
 		break;
@@ -127,7 +127,7 @@ void TextField::setText(const std::string& textIn)
 
 	text.assign(textIn, 0, text.capacity());
 
-	cursorPos = textIn.length();
+	cursorPos = (unsigned short)textIn.length();
 }
 
 void TextField::setText(int value)
